refactor: 물타기 계산기 컴포넌트 폴더 이동
